# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS manager",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/cms-template/public",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/admin/product-comments": {
            "get": {
                "tags": [
                    "Admin Product Comments"
                ],
                "summary": "Get all product comments",
                "operationId": "a0e6557fb6b1d409b28b4220b9b065d8",
                "parameters": [
                    {
                        "name": "is_approved",
                        "in": "query",
                        "description": "Filter by approval status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product comments"
                    }
                }
            }
        },
        "/api/admin/product-comments/{productId}": {
            "get": {
                "tags": [
                    "Admin Product Comments"
                ],
                "summary": "Get product comments by product ID",
                "operationId": "280b4db03e9662a965ee3a3ec84ce96c",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_approved",
                        "in": "query",
                        "description": "Filter by approval status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product comments for the specified product"
                    }
                }
            }
        },
        "/api/admin/product-comments/{id}": {
            "get": {
                "tags": [
                    "Admin Product Comments"
                ],
                "summary": "Show a specific product comment",
                "operationId": "ba6716e41cfd7e9856730a6bd7b75567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product comment details"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin Product Comments"
                ],
                "summary": "Delete a product comment",
                "operationId": "2eb0b6723f0c5c53806cb487747dd9b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result"
                    }
                }
            }
        },
        "/api/admin/product-comments/{id}/approval": {
            "put": {
                "tags": [
                    "Admin Product Comments"
                ],
                "summary": "Set approval status of a comment",
                "operationId": "7feca476b6330f7f0c4ae6bc9ee974ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Approval status",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approval status updated"
                    }
                }
            }
        },
        "/api/admin/product-comments/{commentId}/admin-reply": {
            "post": {
                "tags": [
                    "Admin Product Comments"
                ],
                "summary": "Add an admin reply to a product comment",
                "operationId": "7f24abeee1cd8b0883199ee318639b77",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "The ID of the product comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "description": "The admin's reply text",
                                        "type": "string"
                                    },
                                    "product_comment_rating_id": {
                                        "description": "The ID of the related product comment rating",
                                        "type": "integer"
                                    },
                                    "product_conversation_id": {
                                        "description": "The ID of the related product conversation",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin reply created successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/admin/product-comments/admin-reply/{replyId}": {
            "delete": {
                "tags": [
                    "Admin Product Comments"
                ],
                "summary": "Delete an admin reply",
                "operationId": "3cbfa3c528aad9c37f330e3af9fb7c28",
                "parameters": [
                    {
                        "name": "replyId",
                        "in": "path",
                        "description": "The ID of the admin reply",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin reply deleted successfully"
                    },
                    "404": {
                        "description": "Admin reply not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Authentication using Bearer token (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieSession": {
                "type": "apiKey",
                "description": "Authentication using session cookie (laravel_session)",
                "name": "laravel_session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}